server/requirements.txt
django
djangorestframework
mysql-python